refactor(layout) - rename layout folder
