+++ src/components/sections/experiments/ExperimentsSection/ExperimentsSection.tsx
@@ -4,7 +4,7 @@ import { SectionId } from '@/types/section.types'
 import { experimentsData } from '@/constants/experiments.constants'
 import { i18n } from '@/constants/texts.constants'
 
-import { SectionLayout } from '@/components/layouts'
+import { SectionLayout } from '@/components/layout'
 import { List } from '@/components/UI'
 
 import ExperimentItem from '../ExperimentItem/ExperimentItem'
