+++ src/components/sections/projects/ProjectsSection/ProjectsSection.tsx
@@ -4,7 +4,7 @@ import { SectionId } from '@/types/section.types'
 import { projectsData } from '@/constants/projects.constants'
 import { i18n } from '@/constants/texts.constants'
 
-import { SectionLayout } from '@/components/layouts'
+import { SectionLayout } from '@/components/layout'
 import { List } from '@/components/UI'
 
 import ProjectItem from '../ProjectItem/ProjectItem'
