+++ src/types/navigation.types.ts
